backend/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
decorator==4.4.2
Django==5.0.5
django-cors-headers==4.3.1
django-jazzmin==3.0.0
django-shortuuidfield==0.1.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.4.9
inflection==0.5.1
moviepy==1.0.3
numpy==1.26.4
packaging==24.0
pillow==10.3.0
proglog==0.1.10
PyJWT==2.8.0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
shortuuid==1.0.13
six==1.16.0
sqlparse==0.5.0
tqdm==4.66.4
uritemplate==4.1.1
urllib3==2.2.1
